Run ruff on 3.3

+++ breezy/progress.py
@@ -200,7 +200,7 @@ class DummyProgress:
 def str_tdelta(delt):
     if delt is None:
         return "-:--:--"
-    delt = int(round(delt))
+    delt = round(delt)
     return "%d:%02d:%02d" % (delt / 3600, (delt / 60) % 60, delt % 60)
